include dialogs.h to supress warning.
2001-05-10  Almer S. Tigelaar  <almer@gnome.org>

	* dialog-advanced-filter.c : include dialogs.h
	to supress warning.

	* dialog-cell-sort.c (string_pos_in_list): Cast pointer
	from untyped to char *.

	* dialog-function-wizard.c (formula_guru_set_expr): Fix
	signedness by changing variable decl and some casts.
	(cb_formula_guru_entry_focus_in): Fix signedness by casting,
	we don't want to pollute our data structures with unsigned
	integers.
	(formula_guru_init): Fix signedness by changing decl.

	* dialog-stf-fixed-page.c (fixed_page_autodiscover):
	Fix signedness by changing decl.

	* dialog-stf-preview.c (stf_preview_get_table_pixel_width):
	Fix signedness by changing decl.
	(stf_preview_colformats_clear): Idem.
	(stf_preview_get_column_at_x): Idem.
	(stf_preview_get_column_border_at_x): Idem.
	(stf_preview_get_char_at_x): Idem.

	* dialog-summary.c (dialog_summary_update): Fix signedness by
	cast.

+++ src/dialogs/dialog-summary.c
@@ -104,7 +104,7 @@ dialog_summary_update (WorkbookControlGUI *wbcg, SummaryInfo *sin)
 		return;
 	}
 
-	for (i = 0; i < sizeof(names)/sizeof(char *); i++) {
+	for (i = 0; i < (int) (sizeof(names)/sizeof(char *)); i++) {
 		GtkWidget *entry;
 		entry = glade_xml_get_widget (gui, names[i]);
 		gnome_dialog_editable_enters(GNOME_DIALOG(dia),
